Add advanced filtering
When running tests where multiple request types are used, it if quite difficult to set up proper filtering.

To help this, I made it possible to chain multiple filters and check URL and method.

Furthermore the URL can now be modified.

For future-proofing the Request and Response structs can now be extended with more stuff without breaking compatibility.

If you like this, I can add some tests for the new functionality.