fix: do not write request in cassette response struct (#119)
* fix: do not write request in cassette response struct

Add a `json:"-"` struct tag to the Response.Request field, to make sure
this field is never saved to disk.

This fixes a case where the Response.Request field is populated during a
replay, and then a new track is added, causing the cassette to be
re-saved to disk, including the request field.

* update comment + add TestRecordReplayRecord

* create v17

---------

Co-authored-by: Carl-Magnus Björkell <calle@gibberish.dk>